+++ crd-generator/api/src/test/java/io/fabric8/crd/example/inherited/Base.java
@@ -13,11 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.fabric8.crd;
+package io.fabric8.crd.example.inherited;
 
-/**
- * @author <a href="claprun@redhat.com">Christophe Laprun</a>
- */
-public class BaseSpec {
+import io.fabric8.kubernetes.api.model.Namespaced;
+import io.fabric8.kubernetes.client.CustomResource;
+
+public abstract class Base<Spec extends BaseSpec, Status extends BaseStatus>
+  extends CustomResource<Spec, Status>
+  implements Namespaced {
 
 }