+++ crd-generator/api/src/test/java/io/fabric8/crd/example/inherited/BaseStatus.java
@@ -13,14 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.fabric8.crd.zookeeper.v1alpha1;
+package io.fabric8.crd.example.inherited;
 
-import io.fabric8.kubernetes.model.annotation.SpecReplicas;
+/**
+ * @author <a href="claprun@redhat.com">Christophe Laprun</a>
+ */
+public class BaseStatus {
 
-public class ZookeeperSpec {
-
-    @SpecReplicas
-    private int size;
-    private String version;
-    private boolean ephemeral;
 }