+++ crd-generator/api/src/test/java/io/fabric8/crd/generator/CustomResourceInfoTest.java
@@ -17,6 +17,7 @@ package io.fabric8.crd.generator;
 
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import io.fabric8.kubernetes.api.model.HasMetadata;
 import io.fabric8.kubernetes.api.model.Namespaced;
@@ -67,20 +68,25 @@ public class CustomResourceInfoTest {
   }
 
   @Test
-  void shouldProperlyCreateCustomResourceInfo() {
+  void shouldProperlyCreateCustomResourceInfo() throws Exception {
     CustomResourceInfo info = CustomResourceInfo.fromClass(ClusteredCR.class);
     assertEquals(GROUP, info.group());
     assertEquals(VERSION, info.version());
     assertEquals(Scope.CLUSTER, info.scope());
     assertEquals(ClusteredCR.class.getCanonicalName(), info.crClassName()); // todo: should we actually use the type name here?
-    assertEquals(Spec.class.getTypeName(), info.specClassName().get());
-    assertEquals(Status.class.getTypeName(), info.statusClassName().get());
-    assertEquals(CustomResource.getSingular(ClusteredCR.class), info.singular());
-    assertEquals(CustomResource.getPlural(ClusteredCR.class), info.plural());
+    String specClassName = info.specClassName().get();
+    assertEquals(Spec.class.getCanonicalName(), specClassName);
+    // todo: check that we can load and instantiate class from the returned class name
+    /*Class<?> specClass = Class.forName(specClassName);
+    Object o = specClass.getDeclaredConstructor().newInstance();
+    assertNotNull(o);*/
+    assertEquals(Status.class.getCanonicalName(), info.statusClassName().get());
+    assertEquals(HasMetadata.getSingular(ClusteredCR.class), info.singular());
+    assertEquals(HasMetadata.getPlural(ClusteredCR.class), info.plural());
     assertEquals(CustomResource.getCRDName(ClusteredCR.class), info.crdName());
     assertArrayEquals(CustomResource.getShortNames(ClusteredCR.class), info.shortNames());
-    assertEquals(true, info.served());
-    assertEquals(true, info.storage());
+    assertTrue(info.served());
+    assertTrue(info.storage());
     assertEquals(HasMetadata.getKind(ClusteredCR.class), info.kind());
   }
 }