+++ kubernetes-tests/src/test/java/io/fabric8/crd/generator/inherited/BaseSpec.java
@@ -13,16 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.fabric8.crd.zookeeper.v1;
+package io.fabric8.crd.generator.inherited;
 
-import io.fabric8.kubernetes.model.annotation.SpecReplicas;
-import javax.validation.constraints.NotNull;
+public class BaseSpec {
 
-public class ZookeeperSpec {
-
-    @SpecReplicas
-    private int size;
-    @NotNull
-    private String version;
-    private boolean ephemeral;
 }