+++ kubernetes-tests/src/test/java/io/fabric8/crd/generator/inherited/Child.java
@@ -13,17 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package io.fabric8.crd.generator.inherited;
 
-package io.fabric8.crd.zookeeper.v1;
-
+import io.fabric8.crd.generator.Base;
 import io.fabric8.kubernetes.api.model.Namespaced;
-import io.fabric8.kubernetes.client.CustomResource;
 import io.fabric8.kubernetes.model.annotation.Group;
 import io.fabric8.kubernetes.model.annotation.Version;
 
-@Version("v1")
-@Group("io.zookeeper")
-public class Zookeeper extends CustomResource<ZookeeperSpec, ZookeeperStatus> implements Namespaced {
+@Version("v1alpha1")
+@Group("acme.com")
+public class Child extends Base<ChildSpec, ChildStatus> implements Namespaced {
+
 
 }
-