+++ kubernetes-tests/src/test/java/io/fabric8/crd/generator/zookeeper/v1/ZookeeperSpec.java
@@ -13,16 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.fabric8.crd;
+package io.fabric8.crd.generator.zookeeper.v1;
 
-import io.fabric8.kubernetes.api.model.Namespaced;
-import io.fabric8.kubernetes.client.CustomResource;
+import io.fabric8.kubernetes.model.annotation.SpecReplicas;
+import javax.validation.constraints.NotNull;
 
-/**
- * @author <a href="claprun@redhat.com">Christophe Laprun</a>
- */
-public abstract class Base<Spec extends BaseSpec, Status extends BaseStatus>
-  extends CustomResource<Spec, Status>
-  implements Namespaced {
+public class ZookeeperSpec {
 
+    @SpecReplicas
+    private int size;
+    @NotNull
+    private String version;
+    private boolean ephemeral;
 }