+++ kubernetes-tests/src/test/java/io/fabric8/crd/generator/zookeeper/v1alpha1/Zookeeper.java
@@ -13,11 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.fabric8.crd;
 
-/**
- * @author <a href="claprun@redhat.com">Christophe Laprun</a>
- */
-public class FooSpec extends BaseSpec {
+package io.fabric8.crd.generator.zookeeper.v1alpha1;
 
+import io.fabric8.kubernetes.api.model.Namespaced;
+import io.fabric8.kubernetes.client.CustomResource;
+import io.fabric8.kubernetes.model.annotation.Group;
+import io.fabric8.kubernetes.model.annotation.Version;
+
+@Version(value="v1alpha1", storage=false)
+@Group("io.zookeeper")
+public class Zookeeper extends CustomResource<ZookeeperSpec, ZookeeperStatus> implements Namespaced {
+  private ZookeeperSpec spec;
+  private ZookeeperStatus status;
 }
+