+++ CHANGELOG.md
@@ -3,10 +3,11 @@
 ### 5.4-SNAPSHOT
 
 #### Bugs
-
 * Fix #2989: serialization will generate valid yaml when using subtypes
+* Generating CRDs from the API should now properly work
 
 #### Improvements
+* Fix #2910: Move crd-generator tests from kubernetes-itests to kubernetes-tests
 
 #### Dependency Upgrade
 * Update Kubernetes Model to v1.21.0
@@ -42,7 +43,7 @@
 * Fix #2950: RawCustomResourceOperationsImpl should also work with standard resources
 * Fix #2938: Make it possible to manage Tekton Triggers directly
 * Fix #2921: Kubernetes server mock will generate missing metadata fields
-* Fix #2946: Kubernetes server mock watch will generate initial ADDED events
+* Fix #2946: Kubernetes server mock watch will generate initial `ADDED` events
 * Fix #2925: Add CustomResource.getShortNames(Class) method
 
 #### Dependency Upgrade